pyproject.toml
[project]
name = "agentflow"
version = "0.1.0"
description = "基于FastAPI的MVC架构Agent服务，支持多种检索器配置"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "uvicorn>=0.30.0",
    "pydantic-settings>=2.0.0",
    "langchain-openai>=1.1.5",
    "langchain-huggingface>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-classic>=1.0.0",
    "chromadb>=1.3.7",
    "huggingface-hub>=0.24.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "python-jose>=3.5.0",
    "bcrypt>=5.0.0",
    "python-multipart>=0.0.21",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "email-validator>=2.3.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"  # 指定pip安装时的镜像源，加速依赖下载
extra-index-url = ["https://pypi.org/simple"]  # 添加官方源作为备用

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
